pyproject.toml
[tool.poetry]
name = "home-work-web-13"
version = "0.1.0"
description = ""
authors = ["pentagon846 <pentagon846@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
alembic = "^1.13.1"
bcrypt = "^4.1.2"
cryptography = "^42.0.5"
fastapi-jwt-auth = "^0.5.0"
libgravatar = "^1.0.4"
passlib = "^1.7.4"
uvicorn = "^0.29.0"
psycopg2 = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydentic = "^0.0.1.dev3"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.3"
cloudinary = "^1.39.1"
fastapi-limiter = "^0.1.6"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
aioredis = "^2.0.1"
slowapi = "^0.1.9"
environ = "^1.0"
sphinx = "^7.2.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
